refactor: move cert request (certbot certonly) into separate script
instead of having this done automatically in the entrypoint.

This is so we can have some cleanup (delete directories) before the
request. I don't know a good/clean way to do that automatically (and
every boot).

This introduces an extra step for 1st time setup... but I think it's
less error prone.